Migrate tab selector to header
